requirements.txt
aspose_cells_python==26.1.0
google_api_python_client==2.188.0
google_auth_oauthlib==1.2.4
loguru==0.7.3
Pillow==12.1.0
protobuf==6.33.4
